move database definitions to separate module
